set up dockerfile and modified api to suit google cloud run
